refactor(taxonomy): split library validation out of parsing
Addresses tbittar's review on #278: cross-artifact validation had been put
inside parse_yaml_library, which is the reading/validating mixing that #265
set out to undo.

- New gems_craft/model/validation.py holds check_library_against_taxonomy
  (moved unchanged from taxonomy.py, with its _missing helper) plus
  validate_libraries_against_taxonomy, which absorbs the declared-id and
  missing-taxonomy checks.
- parse_yaml_library goes back to a pure reader: no taxonomy argument, no
  validation. input_libs likewise loses the argument.
- load_study reads input/taxonomy.yml and calls the validation explicitly
  after parsing, alongside consistency_check.
- taxonomy.py no longer imports parsing, so the TYPE_CHECKING guard added to
  break that cycle is gone — the cycle no longer exists.
- Acts on the TODO left by #265: consistency_check moves from
  resolve_components.py to a new gems_craft/study/validation.py; its callers
  are repointed and a now-dead Model import is dropped.

Conformance is enforced when a study is loaded rather than on every parse,
matching how consistency_check and validate_optim_config already work.
