+++ docs/CHANGELOG.md
@@ -9,9 +9,15 @@ All notable changes to GemsPy are documented here.
   *N+1*'s first timestep to block *N*'s **last** timestep — a different absolute timestep. 
 - **linopy upgraded to `>=0.9.0`** - the minimum supported Python version rises
   to **3.11** accordingly (linopy 0.9 requires Python >= 3.11).
-- **Breaking** - parsing a library that declares a `taxonomy` raises `ValueError`
-  if no taxonomy is supplied, or if its id differs from the declared one.
+- **Breaking** - loading a study whose library declares a `taxonomy` raises
+  `ValueError` if no taxonomy is supplied, or if its id differs from the declared
+  one.
 - **Breaking** - `TaxonomyData` renamed to `TaxonomySchema`; no alias kept.
+- **Breaking** - `check_library_against_taxonomy` moved from
+  `gems_craft.model.taxonomy` to the new `gems_craft.model.validation`, and
+  `consistency_check` from `gems_craft.study.resolve_components` to the new
+  `gems_craft.study.validation`. Reading and validating are now separate modules,
+  as in `optim_config/`. No behavior change; import paths only.
 
 ### Added
 - **Integer strategy and thermal heuristics** - components can now set
@@ -30,9 +36,10 @@ All notable changes to GemsPy are documented here.
   model-build time; using them inside constraints, binding-constraints, objective
   contributions, or variable bounds raises a `ValueError`.
 
-- **Taxonomy conformance checked at parse time** - `parse_yaml_library` takes an
-  optional `taxonomy` and checks every library declaring a `taxonomy` field.
-  `load_study` reads it from the optional `input/taxonomy.yml`.
+- **Taxonomy conformance checked when a study is loaded** - `load_study` reads
+  the optional `input/taxonomy.yml` and calls
+  `validate_libraries_against_taxonomy` on every library declaring a `taxonomy`
+  field. `parse_yaml_library` is unchanged and performs no validation.
 
 ### Fixed
 - **Standard library parsing now accepts hybrid port-type fields** -
