+++ tests/e2e/functional/perf_pypsa.py
@@ -12,10 +12,10 @@ from gems_craft.study.data import DataBase
 from gems_craft.study.parsing import parse_yaml_system
 from gems_craft.study.resolve_components import (
     build_data_base,
-    consistency_check,
     resolve_system,
 )
 from gems_craft.study.system import System
+from gems_craft.study.validation import consistency_check
 from gems_runner.simulation import TimeBlock, build_problem
